python-version/YoloVersion/StaticImage/requirements.txt
# YOLO11 Cell Detection Requirements
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
